server/requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17

# ── AI / ML ───────────────────────────────────────────────────────────────────
tensorflow==2.18.0
Pillow==11.0.0
numpy==2.4.2

# ── GIS / Geospatial ──────────────────────────────────────────────────────────
geopandas==1.1.2
osmnx==2.0.7
shapely==2.1.2
pyproj==3.7.2
pyogrio==0.12.1
networkx==3.6.1

# ── Data Processing ───────────────────────────────────────────────────────────
pandas==3.0.0
openpyxl==3.1.5       # Required for reading .xlsx CBS population data
requests==2.32.5

# ── Environment ───────────────────────────────────────────────────────────────
python-dotenv==1.1.0